attached_assets/Ekfv10-main/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "apscheduler>=3.11.0",
    "motor>=3.7.1",
    "paramiko>=3.5.1",
    "py-cord>=2.6.1",
    "python-dotenv>=1.1.0",
]
